% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adjusted.R
\name{Adjust}
\alias{Adjust}
\title{Function for associating race adjusted DNA methylation with matched mRNA expression}
\usage{
Adjust(x, y, alpha, nfolds, clin.col)
}
\arguments{
\item{x}{Processed DNA methylation with covariates. Race must be the final term}

\item{y}{Raw expression count vector}

\item{alpha}{Elastic net mixing penalty. Defaults to 0.5}

\item{nfolds}{the number of folds to use for cross-validation}

\item{clin.col}{the number of clinical covariates included}
}
\value{
returns list of five elements for summarizing model construction
}
\description{
This function allows one to associate DNA methylation with matched mRNA expression
in a multivariate context while adjusting for race. This function assumes
that the variable 'race' is the final column in the matrix.
}
\examples{
data.whole = cbind(matrix(rnorm(500),nrow=50), rbinom(50,1,0.5))
RNA_tmp <- rpois(50,15)
Adjust(data.whole, RNA_tmp, alpha = 0.5, nfolds = 5, clin.col = 1)

}
