% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{annotated_RNA}
\alias{annotated_RNA}
\title{annotated_RNA.}
\format{
A data frame with five variables:
\describe{
\item{\code{ensembl_id}}{Ensebl Gene ID},
\item{\code{chromosome_name}}{chromosome},
\item{\code{start_position}}{Transcription Start Site},
\item{\code{end_position}}{Transcription Stop Site},
\item{\code{hgnc_symbol}}{Gene ID}
}
}
\usage{
annotated_RNA
}
\description{
Annotation for hg38 aligned mRNA.
}
\keyword{datasets}
