% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assoc.R
\name{Assoc}
\alias{Assoc}
\title{Function for associating predicted expression with outcome}
\usage{
Assoc(EXP = EXP, type = c("logistic", "survival", "linear"), clinical = clin)
}
\arguments{
\item{EXP}{Expression predicted using EPI_Predict}

\item{type}{Regression method for associating expression with a phenotype}

\item{clinical}{covariates to be included in the model. Phenotype to test must be in final column (final two for survival)}
}
\value{
Data.frame with the association results for each gene
}
\description{
This function takes the predicted expression from EPI_Predict and tests for
associations with phenotype of interest.
}
\examples{
EXP <- matrix(rpois(500,15),nrow = 50)
rownames(EXP) <- c(1:50)
clin <- as.data.frame(rbinom(50,1,0.5))
Assoc(EXP = EXP, type = 'logistic', clinical = clin)


}
