% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BRCA_Clinical_Test}
\alias{BRCA_Clinical_Test}
\title{BRCA_Clinical_Test.}
\format{
A data frame with two variables:
\describe{
\item{\code{age}}{Age for samples},
\item{\code{race}}{Race of samples},
}
}
\usage{
BRCA_Clinical_Test
}
\description{
Clinical Covariates for BRCA Samples
}
\keyword{datasets}
