% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BRCA_Methy_Test}
\alias{BRCA_Methy_Test}
\title{BRCA_Methy_Test.}
\format{
A matrix with 468 sample and 258 CpG site M-values:
\describe{
\item{\code{BRCA_Methy_Test}}{Ensebl Gene ID},
}
}
\usage{
BRCA_Methy_Test
}
\description{
Test data for vignetta from TCGA-BRCA Methylation.
}
\keyword{datasets}
