% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BRCA_RNA_Test}
\alias{BRCA_RNA_Test}
\title{BRCA_RNA_Test.}
\format{
A data frame with one variables:
\describe{
\item{\code{ENSG00000000003}}{Expression counts for ENSG00000000003},
}
}
\usage{
BRCA_RNA_Test
}
\description{
A data.frame with 468 samples a expression counts for TSPAN6:
}
\keyword{datasets}
