% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Construct.R
\name{Construct}
\alias{Construct}
\title{Function for associating DNA methylation with matched mRNA expression}
\usage{
Construct(
  Object = Object,
  method = c("Adjusted", "Interaction", "Specific"),
  dist = NULL,
  nfolds = NULL,
  impute = TRUE,
  beta = TRUE,
  parallel = FALSE,
  Methylation.Annotation = NULL,
  Gene.Annotation = NULL
)
}
\arguments{
\item{Object}{an object of class Summarized Experiment containing paired
DNA methylation, RNA seq, and Clinical data.}

\item{method}{method for multivariate associations}

\item{dist}{the distance from the TSS & TES in Kb. Defaults to 1,000,000Kb}

\item{nfolds}{the number of folds to use for cross-validation}

\item{impute}{Whether or not to impute the data. Defaults to TRUE}

\item{beta}{Whether methylation matrix is beta-values, defaults to TRUE}

\item{parallel}{whether to run in parallel. Defaults to FALSE}

\item{Methylation.Annotation}{A Granges object annotating the methylation array used}

\item{Gene.Annotation}{A GRanges object containing gene annotation.}
}
\value{
Large list with up to 5 columns
}
\description{
This function allows you to associate DNA methylation with matched mRNA
expression. There are three methods to choose from: 'Adjusted' in which
one can adjust for race, 'Interaction' in which one can enforce a weak hierarchy
and include race interaction terms, and 'Specific' in which one can associate
in a race-specific manner.
}
\examples{
data('Testdat',package = 'EpiXprS')
Methy <- assays(Testdat)$Methyl450k
Methy[1:5,1:5]
Counts <- assays(Testdat)$RNASeqGene
Counts[1:2,1:5]
coldat <- colData(Testdat)
coldat$race <- ifelse(coldat$race == 'white',1,0)
head(coldat)
Construct(x = Methy, y = Counts, clinical = coldat$race,
method ='Interaction', dist = NULL, nfolds = NULL, 
impute = TRUE, beta = TRUE, parallel = FALSE,
Methylation.Annotation = annotation_450k, Gene.Annotation = annotated_RNA)
}
