% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPI.Predict.R
\name{EPI.Predict}
\alias{EPI.Predict}
\title{This function predicts mRNA expression}
\usage{
EPI.Predict(
  Cancer = c("PRAD", "BRCA", "COAD", "KIRP", "KIRC", "HNSC", "LUAD", "UCEC"),
  x = methy,
  clinical = clin
)
}
\arguments{
\item{Cancer}{The type of cancer model to use for expression prediction}

\item{x}{Matrix containing processed methylation M-values}

\item{clinical}{Clinical data}
}
\value{
Matrix of imputed expression values
}
\description{
Using models previously constructed using TCGA reference data, mRNA
expression is predicted using DNA methylation
}
