% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPI.Assoc.R
\name{EPI_Assoc}
\alias{EPI_Assoc}
\title{Function for associating predicted expression with outcome}
\usage{
EPI_Assoc(
  EXP = EXP,
  type = c("logistic", "survival", "linear"),
  clinical = clin
)
}
\arguments{
\item{EXP}{Expression predicted using EPI_Predict}

\item{type}{Regression method for associating expression with a phenotype}

\item{clinical}{covariates to be included in the model. Phenotype to test must be in final column (final two for survival)}
}
\value{
Data.frame with the association results for each gene
}
\description{
This function takes the predicted expression from EPI_Predict and tests for
associations with phenotype of interest.
}
\examples{
EPI.Assoc(EXP = EXP, type = logistic, clinical = cin)

}
