% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPI.Construct.R
\name{EPI_Construct}
\alias{EPI_Construct}
\title{Function for associating DNA methylation with matched mRNA expression}
\usage{
EPI_Construct(
  x = betas,
  y = counts,
  clinical = clin,
  method = c("Adjusted", "Interaction", "Specific"),
  dist = NULL,
  nfolds = NULL,
  impute = TRUE,
  beta = TRUE,
  parallel = FALSE,
  array = "450K"
)
}
\arguments{
\item{x}{matrix of DNA methylation beta values. Can include missing values}

\item{y}{matrix of raw mRNA counts}

\item{clinical}{Data.frame conataing clinical covariates}

\item{method}{method for multivariate associations}

\item{dist}{the distance from the TSS & TES in Kb. Defaults to 1,000,000Kb}

\item{nfolds}{the number of folds to use for cross-validation}

\item{impute}{Whether or not to impute the data. Defaults to TRUE}

\item{beta}{Whether methylation matrix is beta-values, defaults to TRUE}

\item{parallel}{whether to run in parallel. Defaults to FALSE}

\item{array}{Methylation array type, 450k or EPIC, defaults to 450K}
}
\value{
Large list with upto 5 columns
}
\description{
This function allows you to associate DNA methylation with matched mRNA
expression. There are three methods to choose from: 'Adjusted' in which
one can adjust for race, 'Interaction' in which one can enforce a weak hierarchy
and include race interaction terms, and 'Specific' in which one can associate
in a race-specific manner.
}
\examples{
EPI.Construct(x = BRCA_Methy_Test, y = BRCA_RNA_Test,
clinical = BRCA_Clinical_Test, method = 'Interaction', beta = FALSE,
Imputation = FALSE)
}
