% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPI.Predict.R
\name{EPI_Predict}
\alias{EPI_Predict}
\title{This function predicts mRNA expression}
\usage{
EPI_Predict(
  Cancer = c("PRAD", "BRCA", "COAD", "KIRP", "KIRC", "HNSC", "LUAD", "UCEC"),
  x = methy,
  clinical = clin,
  dist = NULL,
  impute = TRUE,
  beta = TRUE
)
}
\arguments{
\item{Cancer}{The type of cancer model to use for expression prediction}

\item{x}{Matrix containing processed methylation beta-values}

\item{clinical}{Clinical data}

\item{dist}{the distance from the TSS & TES in Kb. Defaults to 1,000,000Kb}

\item{impute}{Whether or not to impute the data. Defaults to TRUE}

\item{beta}{Whether methylation matrix is beta-values, defaults to TRUE}
}
\value{
Matrix of imputed expression values
}
\description{
Using models previously constructed using TCGA reference data, mRNA
expression is predicted using DNA methylation
}
\examples{
\dontrun{
EPI_Predict(Cancer = 'PRAD', x = methy, clinical = clin)
}

}
