% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeEpiXprS.R
\name{MakeEpiXprS}
\alias{MakeEpiXprS}
\title{Creates a MultiAssay object containing RNA counts, DNA methylation, and
clinical data.}
\usage{
MakeEpiXprS(
  RNASeqGene = Counts,
  Methyl = Methyl,
  Clinical = Clinical,
  samples = samples
)
}
\arguments{
\item{RNASeqGene}{A matrix of RNA counts}

\item{Methyl}{A matrix of methylation values}

\item{Clinical}{A data.frame of clinical covariates}

\item{samples}{A vector of sample names ordered the same as RNASeqGene and
Methyl450k}
}
\value{
An object of class MultiAssayExperiment containing paired RNA counts
and DNA methylation
}
\description{
Creates a MultiAssay object containing RNA counts, DNA methylation, and
clinical data.
}
\examples{
data('Testdat', package = 'EpiXprS')
Methyl <- assays(Testdat)$Methyl
Counts <- assays(Testdat)$RNASeqGene
Clinical <- colData(Testdat)
samples <- sampleMap(Testdat)$primary[1:100]
Data <- MakeEpiXprS(RNASeqGene = Counts, Methyl = Methyl, 
Clinical = Clinical, samples = samples)
}
