% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_search.R
\name{model_search}
\alias{model_search}
\title{Allows user to search provided models for CpG sites that are associated with
expression of genes in any of the eight cancers}
\usage{
model_search(
  features,
  Cancer = c("PRAD", "BRCA", "COAD", "KIRP", "KIRC", "HNSC", "LUAD", "UCEC")
)
}
\arguments{
\item{features}{CpG sites to be identified in the model of choice}

\item{Cancer}{The model to be searched for associated CpG sites}
}
\value{
data.frame containing all instance of CpG sites in models
}
\description{
Allows user to search provided models for CpG sites that are associated with
expression of genes in any of the eight cancers
}
\examples{
model_search('cg21837192', Cancer = 'PRAD')

}
