% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imputation.R
\name{plogit}
\alias{plogit}
\title{Impute DNA Methylation}
\usage{
plogit(x, min = 0, max = 1)
}
\value{
Matrix of imputed expression
}
\description{
Borrowed from the methyLImp package constructed by Pietro Di Lena
}
