% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{prepare}
\alias{prepare}
\title{Prepare CpG sites and expression for modeling}
\usage{
prepare(
  Counts = Counts,
  Methy = Methy,
  Methylation.Annotation = Methylation.Annotation,
  annotated_RNA = annotated_RNA,
  idx = idx,
  dist
)
}
\arguments{
\item{Counts}{A matrix of RNA counts}

\item{Methy}{A matrix of Methylation M-values}

\item{Methylation.Annotation}{A GRanges object containing CpG site locations.
must have same order as \code{Methy}}

\item{annotated_RNA}{A GRanges object containing RNA annotation. Must have
same order as \code{Counts}}
}
\value{
A list containing the prepared methylation and RNA
}
\description{
Prepare CpG sites and expression for modeling
}
