% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ProcTest}
\alias{ProcTest}
\title{Test data for EpiXprS functionality that has been pre-processed}
\format{
An object of class \code{MultiAssayExperiment} of length 2.
}
\usage{
ProcTest
}
\description{
Test data for EpiXprS from TCGA-BRCA dataset containing the genes A1BG and
A1CF, as well as the CpG sites within 1MB of their transcription start sites
and stop sites.
}
\details{
ProcTest
}
\examples{
data('ProcTest', package = 'EpiXprS')
}
\keyword{data}
