% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Specific.R
\name{Specif}
\alias{Specif}
\title{Function for associating race specific DNA methylation with matched mRNA expression}
\usage{
Specif(x, y, alpha, nfolds, clin.col)
}
\arguments{
\item{x}{Processed DNA methylation with covariates. Race must be the final term}

\item{y}{Raw expression count vector}

\item{alpha}{Elastic net mixing penalty. Defaults to 0.5}

\item{nfolds}{the number of folds to use for cross-validation}

\item{clin.col}{the number of clinical covariates included}
}
\value{
returns list of five elements for summarizing model construction
}
\description{
This function allows one to associate DNA methylation with matched mRNA expression
in a multivariate context in race specific data.
}
