% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TestData}
\alias{TestData}
\title{Test data for EpiXprS functionality}
\format{
An object of class \code{SummarizedExperiment} with 258 rows and 468 columns.
}
\usage{
TestData
}
\description{
Test data consisting of expression counts for TSPAN6 from the TCGA-BRCA
study as well as methylation M-values and clinical data for the patients.
}
\keyword{data}
