<<<<<<< HEAD:man/Predict.Rd
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict.R
\name{Predict}
\alias{Predict}
\title{This function predicts mRNA expression}
\usage{
Predict(
  Cancer = object,
  Methy = methy,
  clinical = clin,
  dist = NULL,
  impute = TRUE,
  beta = TRUE
)
}
\arguments{
\item{Cancer}{ExperimentHub object cancer model}

\item{Methy}{Matrix containing processed methylation beta-values}

\item{clinical}{Clinical data}

\item{dist}{the distance from the TSS & TES in Kb. Defaults to 1,000,000Kb}

\item{impute}{Whether or not to impute the data. Defaults to TRUE}

\item{beta}{Whether methylation matrix is beta-values, defaults to TRUE}
}
\value{
Matrix of imputed expression values
}
\description{
Using models previously constructed using TCGA reference data, mRNA
expression is predicted using DNA methylation
}
\examples{
data(BRCA_Methy_Test)
data(BRCA_Clinical_Test)
colnames(BRCA_Clinical_Test) <- c('age','race_list')
BRCA_Clinical_Test <- as.data.frame(BRCA_Clinical_Test)
Predict(Cancer = 'BRCA', Methy = BRCA_Methy_Test, clinical = BRCA_Clinical_Test,
impute = FALSE, beta = FALSE)


}
=======
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_Predict.R
\name{methyl_Predict}
\alias{methyl_Predict}
\title{This function predicts mRNA expression}
\usage{
methyl_Predict(
  Cancer = c("PRAD", "BRCA", "COAD", "KIRP", "KIRC", "HNSC", "LUAD", "UCEC"),
  Methy = methy,
  clinical = clin,
  dist = NULL,
  impute = TRUE,
  beta = TRUE
)
}
\arguments{
\item{Cancer}{The type of cancer model to use for expression prediction}

\item{Methy}{Matrix containing processed methylation beta-values}

\item{clinical}{Clinical data}

\item{dist}{the distance from the TSS & TES in Kb. Defaults to 1,000,000Kb}

\item{impute}{Whether or not to impute the data. Defaults to TRUE}

\item{beta}{Whether methylation matrix is beta-values, defaults to TRUE}
}
\value{
Matrix of imputed expression values
}
\description{
Using models previously constructed using TCGA reference data, mRNA
expression is predicted using DNA methylation
}
\examples{
data(BRCA_Methy_Test)
data(BRCA_Clinical_Test)
colnames(BRCA_Clinical_Test) <- c('age','race_list')
BRCA_Clinical_Test <- as.data.frame(BRCA_Clinical_Test)
methyl_Predict(Cancer = 'BRCA', Methy = BRCA_Methy_Test, clinical = BRCA_Clinical_Test,
impute = FALSE, beta = FALSE)


}
>>>>>>> upstream/master:man/methyl_Predict.Rd
